Create a separate server config for Apollo tests
