+++ src/admin/server.ts
@@ -24,5 +24,5 @@ export const server = new ApolloServer({
     s3,
   },
   uploads: false,
-  // playground: true,
+  playground: true,
 });
